add getMappings.js utility
