+++ lib/convert/MD.js
@@ -1,11 +1,7 @@
-import createCSVStream from 'csv-parse';
-import ndjson          from 'ndjson';
-import setKeys         from '../utilities/setKeys.js';
-
-import {
-  createReadStream,
-  createWriteStream,
-} from 'fs';
+import { createWriteStream } from 'fs';
+import ndjson                from 'ndjson';
+import readTSV               from '../utilities/readTSV.js';
+import setKeys               from '../utilities/setKeys.js';
 
 /**
  * Extracts certain kinds of information from the definition and updates the Sense and Lexeme objects as appropriate.
@@ -80,34 +76,6 @@ function convertRecord({
 }
 
 /**
- * Reads the CSV file into an Array of JSON objects
- * @param  {String}         inputPath The path to the MD CSV file.
- * @return {Promise<Array>}           Resolves to an Array of Objects representing CSV entries.
- */
-async function readCSV(inputPath) {
-
-  const csvOptions = {
-    columns:   true,
-    delimiter: '\t',
-    relax:     true,
-  };
-
-  const readStream = createReadStream(inputPath);
-  const csvStream  = createCSVStream(csvOptions);
-
-  readStream.pipe(csvStream);
-
-  const records = [];
-
-  for await (const record of csvStream) {
-    records.push(record);
-  }
-
-  return records;
-
-}
-
-/**
  * Splits a string into senses based on sense numbers (1., 2. etc.).
  * @param  {String} definition The string to split.
  * @return {Array}             Returns an array of definitions, even when only one sense/definition is present.
@@ -162,7 +130,7 @@ export default async function convertCW(inputPath, outputPath) {
     throw new Error('Please provide the path to the Maskwacîs database as the first argument.');
   }
 
-  const records = await readCSV(inputPath);
+  const records = await readTSV(inputPath);
   const entries = records.map(convertRecord);
 
   setKeys(entries);
